a few bugs to fix
