+++ packages/components/src/Form/Inputs/InputFilters/FilterEditor.test.tsx
@@ -23,40 +23,28 @@
  SOFTWARE.
 
  */
-import React, { FC } from 'react'
-import { ButtonGroup, ButtonItem } from '../../../Button'
-import { PopoverContent } from '../../../Popover'
-import { SpaceVertical } from '../../../Layout/Space'
-import { FieldFilter } from './InputFilters'
 
-interface DraftFilterProps {
-  draft?: FieldFilter
-  options: string[]
-  onClick: (field: FieldFilter, value: string) => void
-}
-export const DraftFilter: FC<DraftFilterProps> = ({
-  draft,
-  onClick,
-  options,
-}) => {
-  if (!draft) return null
+import React from 'react'
+import { renderWithTheme } from '@looker/components-test-utils'
+import { fireEvent } from '@testing-library/react'
+import { FilterEditor } from './FilterEditor'
 
-  return (
-    <PopoverContent p="large" width="360px">
-      <SpaceVertical>
-        <ButtonGroup>
-          {options.map((option, index) => {
-            const handleDraft = () => {
-              onClick(draft, option)
-            }
-            return (
-              <ButtonItem key={index} onClick={handleDraft}>
-                {option}
-              </ButtonItem>
-            )
-          })}
-        </ButtonGroup>
-      </SpaceVertical>
-    </PopoverContent>
-  )
-}
+describe('InputFilters', () => {
+  const draft = { field: 'Name', label: 'name' }
+  const options = ['suggestion 1', 'suggestion 2', 'suggestion 3']
+  const handleDraft = jest.fn()
+
+  test('renders DraftFilter', () => {
+    const { queryByText } = renderWithTheme(
+      <FilterEditor draft={draft} options={options} onClick={handleDraft} />
+    )
+    const draftOption = queryByText('suggestion 1')
+    draftOption && fireEvent.click(draftOption)
+
+    expect(draftOption).toBeInTheDocument()
+    expect(handleDraft).toBeCalled()
+
+    // Close popover to silence act() warning
+    fireEvent.click(document)
+  })
+})
